LayoutTests/imported/w3c/web-platform-tests/wai-aria/role/table-roles-expected.txt

Tests table and related roles.

x
x

x x
x x
x
x x x
x x x
x
x
x x
x x
x x x
x x x x
x x
x x

PASS div role is caption (in div with table role)
PASS orphan p role is caption
PASS span role is cell (in div with row role, in div with rowgroup role, in div with table role)
PASS orphan span role is cell
PASS span role is columnheader (in div with row role, in div with rowgroup role, in div with table role)
PASS div role is row (in div with rowgroup role, in div with table role)
PASS div role is rowgroup (in div with table role)
PASS role is rowheader (in div with row role, in div with rowgroup role, in div with table role)
PASS div role is table
